refactor: adds file for flagbutton
